Added mdb property to request objects.
Extended pyramid's request object adding a reifyed property called "mdb"
which represent the connection to the mongodb database as configured
in the ini file.
Every request has its own connection which is closed at the end of the
request itself. Multiple accesses to the mdb property inside a single
request keep returning the same object.

Added configuration options to the development/production ini as an
example on how to configure.